simplified the Apache configuration: It is no longer needed to configure a database query. At the same time it's less error-prone and avoids trouble if one forgets to update the query, when the database schema changes.
This can become a real problem if the disk fills up due to the Apache
error_log being spammed with errors or warnings due to an outdated query -
and in fact this happened twice to me, so much for the motivation.

The change required moving the preparation of SQL prepared statements into a
post_config module hook. Previously, we did this during configuration processing,
each time we encountered a ZrkadloDBDQuery directive, but since those are not
necessarily present anymore we need to do it later.
The query is now stored as server wide setting, not in directory context.