Parallelizing mirror probing.
For this new feature, the Python module "processing" or "multiprocessing"
needs to be installed.  If none of them is found, the fallback behaviour is to
probe serially, as before.

This affects the "probefile" and the "file ls --probe" commands. It also
affects the scanner, in that it parallelizes the pre-scan check for existing
directories, if the scanner is started to scan a subdirectory only.
The mirrorprobe is not affected by this changes; it runs in threads.

Other user visible changes:
- mb probefile: 
  - implemented downloading and displaying the content.
  - an --urls switch was added, to select the kind of URLs to be probed.
    --urls=scan probes the URLs that would be used in scanning. 
    --urls=http probes the (HTTP) base URLs used in redirection.
    --urls=all probes all registered URLs.
  - proxy envvars are unset before probing.

Internal changes:
- mb.testmirror module:
  - req() is basically replaced by the probe() function, which takes Sample
    instances as arguments. A req() function (which wraps around probe()) is
    retained for backwards compatibility, although it's no longer in use by us.
- mb.util: 
  - a class named Sample was added which can be used to schedule URLs to probe,
    as well as collect the results.
    A reason for this is that the multiprocessing module need to pass objects
    that can be pickled, which isn't the case for SQLObjects' native objects.
    In addition, the class is easily extensible and much easier to use.
  - the time format output by the timer was slightly changed.