mb new: save some dns lookups and support querying geoip using ipv6
Instead of passing the hostname to the various geolocation lookup functions
and have the tool do a ip resolution over and over, we just get the info upfront
and pass it down to the tools.

This also allows us to decide if we shall call geoiplookup or geoiplookup6, in case
a server only supplies ipv6 addresses. Then geoiplookup could not possibly discover
anything