Refactor ASN and GeoIP handling (#11)
GeoIP data files are deprecated in favor or the new maxminddb format.

During the port a lot of code duplication was eliminated by moving the
functionality of asn.py and geoip.py into a MirrorBrainHost class
(formerly IPAddress/MbIpAddress). This also reduced the complexity of
the code that uses the functionality.

Instead of probing for different database files we now allow configuring
the database files that should be used. The default files are:

```
/var/lib/GeoIP/GeoLite2-ASN.mmdb
/var/lib/GeoIP/GeoLite2-City.mmdb
```

Other paths can be configured in /etc/mirrorbrain.conf:

```
[general]
maxmind_asn_db = /path/to/db
maxmind_city_db = /path/to/db
```

mod_asn support is not fully removed yet but already planned to be
removed. We can get the same functionality with the ASN maxminddb
without a DB roundtrip.